+++ main1026.cpp
@@ -175,7 +175,7 @@ else if (tdB_3_4 != 66308774)
 int j = 0;
 while (j < 8)
 {
-std::cout << v773gK_d8 << bXi_CG(857210426, bGG98, -1228187032, (unsigned short)-972568506) << bXi_CG(579173483, -306862204, tdB_3_4, (unsigned short)-218680412) << vCB_ << bXi_CG(ZQU_, bGG98, 1893340332, -1048578428) << std::endl;
+std::cout << v773gK_d8 << bXi_CG(857210426, bGG98, -1228187032, (unsigned short)-972568506) << bXi_CG(579173483, -306862204, tdB_3_4, (unsigned short)-218680412) << vCB_ << bXi_CG(ZQU_, bGG98, 1893340332, (unsigned short)-1048578428) << std::endl;
 j++;
 }
 }
