+++ main1094.cpp
@@ -13,7 +13,7 @@ int main(int argc, char** argv) {
     unsigned short    kp81n6p1,    dcRgcRTT,    q_5Lt8qGo,    Ql3zoZ_r,    du0a6nlQz0,    CpXi,    K8xB_J52pA,    J9wY,    w9U7jov81,    Edc4,    MS5U5r,    y9V84,    XQdx3,    o8a5KTS,    ZhG19wV7;
     kp81n6p1    =    (unsigned short)972911212;
     dcRgcRTT    =    (unsigned short)-1363668378;
-    q_5Lt8qGo    =    -1631699966;
+    q_5Lt8qGo    =    (unsigned short)-1631699966;
     Ql3zoZ_r    =    1819448592;
     du0a6nlQz0    =    -1799762045;
     CpXi    =    1827354453;
@@ -23,7 +23,7 @@ int main(int argc, char** argv) {
     Edc4    =    1979779421;
     MS5U5r    =    (unsigned short)-1335124701;
     y9V84    =    (unsigned short)-1415065052;
-    XQdx3    =    -11176939;
+    XQdx3    =    (unsigned short)-11176939;
     o8a5KTS    =    -1748259546;
     ZhG19wV7    =    -1902293527;
     unsigned long long    fYvak3tHl,    vlpkc7a_,    d_CH75_ZrO,    d3aW54TCl;
